+++ src/_components/MaeilWiki/Home/WikiList/index.tsx
@@ -1,21 +1,10 @@
-import Image from 'next/image';
 import React from 'react';
-import ChatIcon from '@/_assets/icons/chatIcon.svg';
-import {
-  chatIcon,
-  commentCountInfo,
-  ownerImage,
-  ownerInfo,
-  wikiItem,
-  wikiList,
-  wikiQuestion,
-  wikiQuestionOrder,
-  wikiQuestionText,
-  wikiSubInfo,
-} from './wikiList.css';
+import { wikiList } from './wikiList.css';
 import mockWikiList from '../mockWikiList.json';
+import Paginator from '@/_components/common/Paginator';
+import WikiItem from './WikiItem';
 
-interface Member {
+export interface Member {
   profileImage: string;
   name: string;
   github: string;
@@ -38,7 +27,7 @@ interface Wiki {
   ];
 }
 
-type WikiListItem = Omit<Wiki, 'questionDetail' | 'comments'> & { commentCount: number };
+export type WikiListItem = Omit<Wiki, 'questionDetail' | 'comments'> & { commentCount: number };
 
 const MOCK_WIKI_LIST = mockWikiList as {
   isLastPage: boolean;
@@ -54,48 +43,7 @@ export default function WikiList() {
           <WikiItem key={wiki.id} wiki={wiki} order={index + 1} />
         ))}
       </ol>
-    </div>
-  );
-}
-
-function WikiItem({ wiki, order }: { wiki: WikiListItem; order: number }) {
-  const { question, owner, commentCount } = wiki;
-
-  const displayingCommentCount = commentCount || '미답변';
-
-  return (
-    <li className={wikiItem}>
-      <div className={wikiQuestion}>
-        <div className={wikiQuestionOrder}>
-          <span>{order}</span>
-        </div>
-        <p className={wikiQuestionText}>{question}</p>
-      </div>
-      <div className={wikiSubInfo}>
-        <OwnerInfo owner={owner} />
-        <div className={commentCountInfo}>
-          <ChatIcon className={chatIcon} /> {displayingCommentCount}
-        </div>
-      </div>
-    </li>
-  );
-}
-
-function OwnerInfo({ owner }: { owner?: Member }) {
-  if (!owner) {
-    return <div>질문 제공자: 익명</div>;
-  }
-
-  return (
-    <div className={ownerInfo}>
-      <p>질문 제공자: {owner?.name}</p>
-      <Image
-        width={17}
-        height={17}
-        className={ownerImage}
-        src={owner.profileImage}
-        alt={`작성자 ${owner.name}님의 프로필 이미지`}
-      />
+      <Paginator currentPage={1} lastPage={11} />
     </div>
   );
 }
